Extremely early stages of engine + fantasy xcom style game
Building out the engine as pieces become necessary for the creation of a high fantasy xcom-like game. Basic stuff like component infrastructure, ECS system, core opengl rendering functionality, vector math are all in place, but it's all very bare bones at the moment. The "game" is just an isometric map with a single unit on it that can be moved around by clicking. Still, many of the big pieces (responding to events, drawing stuff, updating game state) are real enough to use. Just started out implementing animations in the context of the game, as animations are a recurring challenge and can substantially impact the design of the data management itself. This time around I don't think we're going to go for any fancy ledger based system, just boring old raw mutable data with the rendering layer reponsible for looking at what animations are currently in-progress and adapting the rendering accordingly. Animation may be a hard problem, but it's maybe not a hard enough problem to justify making every single other part of the engine more complicated to support.

Also includes a license making clear (hopefully) that the code is all happily MIT licensed but the assets are not since many of them are purchased or adapted from purchased assets.