More foundational work on hfcom, pathfinding, data, etc
Starting to fill out all of the data representation of things like character classes, items, actions and the like. Still plenty to do, but starting to get the bare bones in place.
The main core work is primarily around making Config pretty printing work for all of the game data. That should make it much easier to get a good introspection on game state.