Vision / shadowcasting, hfcom objects
The big chunk of work here is implementing shadowcasting of a couple different varieties. The old homegrown algorithm is included, though that's really much more applicable to actual light, since it bleeds around edges. For vision this now includes restrictive precise angle shadowcasting, though my implementation is actually a bit different than the canonical. Specifically, I based it on the actual angles of the geometry involved (an occupied square is treated as a literal square) rather than treating all cells in an arc as having equal portions of that arc. I've also extended it into 3d, and the results are honestly pretty decent and the speed is manageable though there's almost certainly still room for improvement. Supports non 1:1 resolution so you can effectively do supersampling of a kind, or work with non-cubic terrains (my actual use case here).

On the hfcom side mostly the changes are just around adding support for objects, that is, things in the world that are not terrain but interact with vision and movement. The changes required for that are much less substantial than the shadowcasting / vision stuff though.